mapped over the results array and display as seperate p elements
